Move option validation to it's own validator
