+++ app/validators/answers/__init__.py
@@ -1,6 +1,7 @@
 from app.validators.answers.answer_validator import AnswerValidator
 from app.validators.answers.date_answer_validator import DateAnswerValidator
 from app.validators.answers.number_answer_validator import NumberAnswerValidator
+from app.validators.answers.option_answer_validator import OptionAnswerValidator
 from app.validators.answers.text_field_answer_validator import TextFieldAnswerValidator
 
 
@@ -11,5 +12,7 @@ def get_answer_validator(answer, questionnaire_schema):
         "Number": NumberAnswerValidator,
         "Currency": NumberAnswerValidator,
         "Percentage": NumberAnswerValidator,
+        "Dropdown": OptionAnswerValidator,
+        "Radio": OptionAnswerValidator,
     }
     return validators.get(answer["type"], AnswerValidator)(answer, questionnaire_schema)
